requirements.txt
fastapi~=0.115.6
pymongo~=4.10.1
pandas~=2.2.3
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1
psycopg2~=2.9.10
openai~=0.28.0
pdfplumber~=0.11.4